+++ 3a-lista/Exercicios/src/exercicio1/Retangulo.java
@@ -22,8 +22,8 @@ public class Retangulo implements CalculoArea {
     }
 
     @Override
-    public float area() {
-        float area;
+    public double area() {
+        double area;
         area = comprimento * altura;
         
         return area;
